requirements.txt
beautifulsoup4==4.9.1
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
flaskwebgui==0.1.9
future==0.18.2
google==2.0.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psutil==5.7.0
python-whois==0.7.3
soupsieve==2.0.1
SQLAlchemy==1.3.17
visitor==0.1.3
Werkzeug==1.0.1
whichcraft==0.6.1
WTForms==2.3.1
